Remove ArchiveEntryDocument and clean up ArchiveWalker.
